requirements.txt
stomp.py
crawl4ai
python-dotenv
google-genai
openai
readabilipy
psycopg[binary]
